First attempt at proper energy strategy: calculate power in supercap and make some guesses as to the state of the secondary battery based on the value of VBatOK.  Report an error if no ack is received for a message that requires an ack (which they all do at the moment).  VBAT_OK_GOOD_THRESHOLD_MV becomes 3.8V from 3.6V, cos I think the modem needs the oomph.  ZOEM8_POWER_ACTIVE_NW reduced to a more realistic number.  Change debugGetStackMinLeft() to go through all the individual stacks and find the smallest number remaining, rather more useful than the previous aggregate number. Also print out the min left stack for all tasks I generate at the end of the task run so that I can then so who is left with what.  Add a minimal implementation of atoi() since calling the actual atoi() requires a change to a configuration value (OS_THREAD_LIBSPACE_NUM) in mbed-os.  Add THIS_STACK_MIN_LEFT, EVENT_NOT_ENOUGH_POWER_TO_RUN_PROCESSOR, EVENT_PROCESSOR_RUNNING, EVENT_RETURN_TO_SLEEP, ENERGY_AVAILABLE_UWH, EVENT_IMEI_ENDING and rename EVENT_WAITING_POWER to EVENT_WAITING_ENERGY; APPLICATION_LOG_VERSION becomes 7.  Remove some remaining PRINTF()s which aren't required now that we have the right log points and the option to print them immediately.  Use #pragma to stop "assignment in condition" warning that ARMCC insists on issuing whatever else I do to try to stop it.  uint64_t changed to unsigned long long int for consistency with all the other sloppy-old C types.

+++ source/actions/act_modem.h
@@ -115,8 +115,8 @@ bool modemIsR2();
  * @return                 the energy consumed in nanoWatt hours,
  *                         limiting at 0xFFFFFFFF on overflow.
  */
-uint64_t modemEnergyNWH(unsigned int idleTimeSeconds,
-                        unsigned int bytesTransmitted);
+unsigned long long int modemEnergyNWH(unsigned int idleTimeSeconds,
+                                      unsigned int bytesTransmitted);
 
 #endif // _ACT_MODEM_H_
